[Hackability Refactor] Move Browser/OpenAI/Server under torchchat usages (#1052)
* Initial Move of Browser/OpenAI/Server under torchchat usages

* Fix swapped init files
