+++ README.md
@@ -250,7 +250,7 @@ First, follow the steps in the Server section above to start a local server. The
 [skip default]: begin
 
 ```
-streamlit run browser/browser.py
+streamlit run torchchat/usages/browser.py
 ```
 
 Use the "Max Response Tokens" slider to limit the maximum number of tokens generated by the model for each response. Click the "Reset Chat" button to remove the message history and start a fresh chat.
