+++ requirements.txt
@@ -17,6 +17,7 @@ gguf
 lm-eval==0.4.2
 blobfile
 tomli >= 1.1.0 ; python_version < "3.11"
+openai
 
 # Build tools
 wheel
